1. Minimized and updated prompt. 2. Updated readme 3. Moved prompt to prompy.py
