+++ ui/streamlit_app.py
@@ -8,10 +8,10 @@ except:
     API_URL = "http://localhost:8000"
 
 st.set_page_config(
-    page_title="Durable Intelligent Loan Underwriter Copilot",
+    page_title="Durable Intelligent Loan Underwriter",
     page_icon="",
 )
-st.title("Durable Intelligent Loan Underwriter Copilot — Demo")
+st.title("Durable Intelligent Loan Underwriter — Demo")
 
 tab = st.tabs(["Submit", "Review", "Workflows"])
 st.set_page_config(layout="wide")
